Added Controller and Comments
I added the default controller which is a template from the back-end meeting and I commented both controllers so that we understood them more
